+++ src/modules/api_act.js
@@ -2,7 +2,6 @@ var logger =require("./logger");
 var cloud = require("./waitForCloud");
 var fhparams = require("./fhparams");
 var ajax = require("./ajax");
-var JSON = require("JSON");
 var handleError = require("./handleError");
 var appProps = require("./appProps");
 
@@ -13,7 +12,7 @@ function doActCall(opts, success, fail){
   params = fhparams.addFHParams(params);
   return ajax({
     "url": url,
-    "tryJSONP": true,
+    "tryJSONP": typeof Titanium === 'undefined',
     "type": "POST",
     "dataType": "json",
     "data": JSON.stringify(params),
@@ -23,7 +22,7 @@ function doActCall(opts, success, fail){
     "error": function(req, statusText, error){
       return handleError(fail, req, statusText, error);
     }
-  })
+  });
 }
 
 module.exports = function(opts, success, fail){
@@ -45,5 +44,5 @@ module.exports = function(opts, success, fail){
     } else {
       doActCall(opts, success, fail);
     }
-  })
-}
+  });
+};