+++ src/modules/api_cloud.js
@@ -2,7 +2,6 @@ var logger =require("./logger");
 var cloud = require("./waitForCloud");
 var fhparams = require("./fhparams");
 var ajax = require("./ajax");
-var JSON = require("JSON");
 var handleError = require("./handleError");
 var appProps = require("./appProps");
 
@@ -30,7 +29,7 @@ function doCloudCall(opts, success, fail){
     "error": function(req, statusText, error){
       return handleError(fail, req, statusText, error);
     }
-  })
+  });
 }
 
 module.exports = function(opts, success, fail){
@@ -48,5 +47,5 @@ module.exports = function(opts, success, fail){
     } else {
       doCloudCall(opts, success, fail);
     }
-  })
-}
+  });
+};