+++ src/modules/handleError.js
@@ -1,5 +1,3 @@
-var JSON = require("JSON");
-
 module.exports = function(fail, req, resStatus, error){
   var errraw;
   var statusCode = 0;
@@ -7,7 +5,7 @@ module.exports = function(fail, req, resStatus, error){
     try{
       statusCode = req.status;
       var res = JSON.parse(req.responseText);
-      errraw = res.error || res.msg;
+      errraw = res.error || res.msg || res;
       if (errraw instanceof Array) {
         errraw = errraw.join('\n');
       }